Fixing bugs

+++ api/v3/modules/generate_retro_website.py
@@ -1,4 +1,4 @@
-import openai, requests, bs4, io, json
+import openai, requests, bs4, io, re
 
 
 class NamedBytesIO(io.BytesIO):
@@ -7,13 +7,13 @@ class NamedBytesIO(io.BytesIO):
         self.name = name
 
 
-def generate_retro_website(url, api_key, assistant_id):
+def generate_retro_website(url, api_key, assistant_id, topP, temperature):
     response = requests.get(url)
     response.raise_for_status()
     soup = bs4.BeautifulSoup(response.text, 'html.parser')
     html = str(soup)
 
-    path = "api/v3/cache/file.txt"
+    path = "api/cache/file.txt"
 
     with open(path, 'w') as file:
         file.write(html)
@@ -21,7 +21,11 @@ def generate_retro_website(url, api_key, assistant_id):
     client = openai.OpenAI(api_key=api_key)
 
     file = client.files.create(file=open(path, "rb"), purpose='assistants')
-    assistant = client.beta.assistants.retrieve(assistant_id=assistant_id)
+    assistant = client.beta.assistants.update(
+        assistant_id=assistant_id,
+        top_p=topP,
+        temperature=temperature
+    )
 
     thread = client.beta.threads.create(
         messages=[
@@ -46,12 +50,38 @@ def generate_retro_website(url, api_key, assistant_id):
     if run.status == 'completed':
         messages = client.beta.threads.messages.list(thread_id=thread.id)
 
-        html_file_id = messages.data[0].attachments[0].file_id
-        content = client.files.with_raw_response.retrieve_content(html_file_id).content.decode('utf-8')
+        print(messages)
 
-        client.files.delete(file.id)
-        client.files.delete(html_file_id)
+        if len(messages.data[0].attachments):
+            html_file_id = messages.data[0].attachments[0].file_id
+            css = client.files.with_raw_response.retrieve_content(html_file_id).content.decode('utf-8')
+            try:
+                client.files.delete(file.id)
+                client.files.delete(html_file_id)
+            except:
+                pass
+        else:
+            css = messages.data[0].content[0].text.value
+            css = css[css.find("```css") + 6:css.rfind("```")]
 
-        return content
+        css = css.replace(";", " !important;")
+        css = css.replace("```css", "")
+        css = css.replace("```", "")
+        css = css.replace("<style>", "")
+        css = css.replace("</style>", "")
+        css = "<style>*{border-radius: 0 !important;}" + f"\n{css}\n</style>"
+
+        body_opening_tag_pattern = r"(<body[^>]*>)"
+
+        html = re.sub(
+            body_opening_tag_pattern,
+            r"\1\n" + css,
+            html,
+            count=1
+        )
+
+        return html
     else:
+        print(run)
+        print(run.status)
         raise RuntimeError("Schema was not generated")
